python/di_to_cu_migration_tool/requirements.txt
# Dependencies for the DI to CU Migration OSS Tool
azure-ai-contentunderstanding==1.0.0
azure-core==1.34.0
azure-identity==1.20.0
azure-storage-blob==12.25.1
dotenv==0.9.9
pytest==8.3.5
python-dateutil==2.9.0post0
rich==13.9.4
typer==0.15.1
